sked changes

+++ capstone/src/components/Professor.jsx
@@ -47,13 +47,18 @@ function Professor() {
   const handleCloseModals = () => {
     setShowModal(false);
   };
+
   const handleDeleteBtnClick = async (userId) => {
-    try {
-      await deleteDoc(doc(db, 'professor', userId));
-      // After successfully deleting the document, update the state to remove the deleted user from the table
-      setUsers((prevUsers) => prevUsers.filter((user) => user.id !== userId));
-    } catch (error) {
-      console.error("Error deleting document: ", error);
+    const confirmDelete = window.confirm("Are you sure you want to delete this User?");
+    
+    if (confirmDelete) {
+      try {
+        await deleteDoc(doc(db, "professor", userId));
+        // After successfully deleting the document, update the state to remove the deleted user from the table
+        setUsers((prevUsers) => prevUsers.filter((user) => user.id !== userId));
+      } catch (error) {
+        console.error("Error deleting document: ", error);
+      }
     }
   };
